Move fallback expansion to separate module
